all: allow customization via tags
The parser now interprets an easyjson struct comment as a list of tags.
For example:

    //easyjson:tag1,tag2
    type T struct { ... }

These tags can be used to customize code generation. For now, easyjson
understands the following tags: "nomarshal", "nounmarshal", and "json".
The "nomarshal" and "nounmarshal" tags tell easyjson to not generate
marshaling or unmarshaling code. The "json" tag behaves as before (it
tells easyjson to generate code for this struct).